### 2019-03-06 #1302 @whedon
[PRE REVIEW]: ACHR.cu: GPU sampling of metabolic networks.
**Submitting author:** @marouenbg (<a href="http://orcid.org/0000-0001-5934-966X">Marouen Ben Guebila</a>)
**Repository:** <a href="https://github.com/marouenbg/ACHR.cu" target ="_blank">https://github.com/marouenbg/ACHR.cu</a>
**Version:** v0.1
