### 2019-03-06 63aa671
merge columns
