### 2019-03-06 0c6a690
fixed table legend
