+++ README.md
@@ -54,12 +54,14 @@ This command allows to generate warmup points given by the user in runtime of th
 
 2. The actual sampling starting from the previously generated warmup points.
 Quick install first.
+
 ```
 cd ACHRcu
 source ./install.sh
 make
 ```
 Then perform the sampling.
+
 `./ACHR model.mps warmuppoints.csv nFiles nPoints nSteps`
 
 This command allows to generate the actual sampling points starting from the previously generated sampling points stored in `warmuppoints.csv` to generate a total of `nFiles*nPoints` with each point
