### 2019-03-12 #1302 @lpantano
Hi @marouenbg,

I will lead the review process of this tool. Thank you for submitting. I am looking for possible reviewers, meanwhile, could you add the DOI links to the reference, please? It has to end up something like this: https://github.com/openjournals/joss-papers/blob/joss.01299/joss.01299/10.21105.joss.01299.pdf

Thanks!
