### 2019-03-12 313cbd5
added doi references
