+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={e86587},
   year={2014},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0086587},
 }
 
 @article{guebila2018dynamic,
@@ -15,7 +16,8 @@
   journal={bioRxiv},
   pages={440701},
   year={2018},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={https://doi.org/10.1101/440701},
 }
 
 @inproceedings{nickolls2008scalable,
@@ -24,7 +26,8 @@
   booktitle={2008 IEEE Hot Chips 20 Symposium (HCS)},
   pages={40--53},
   year={2008},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1145/1365490.1365500},
 }
 
 @article{kaufman1998direction,
@@ -35,7 +38,8 @@
   number={1},
   pages={84--95},
   year={1998},
-  publisher={INFORMS}
+  publisher={INFORMS},
+  doi={10.1287/opre.46.1.84},
 }
 
 @InProceedings{VBCG_HPCS14,
@@ -47,6 +51,7 @@
    month =        {July},
    address =      {Bologna, Italy},
    publisher =    {IEEE},
+   doi={doi.org/10.1109/hpcsim.2014.6903792},
 }
 
 @article{heirendt2019creation,
@@ -56,7 +61,8 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   journal={Nature protocols},
   pages={1},
   year={2019},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41596-018-0098-2},
 }
 
 @article{o2015using,
@@ -67,7 +73,8 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   number={5},
   pages={971--987},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={doi: 10.1016/j.cell.2015.05.019},
 }
 
 @article{mahadevan2003effects,
@@ -78,7 +85,8 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   number={4},
   pages={264--276},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ymben.2003.09.002},
 }
 
 @article{dagum1998openmp,
@@ -88,5 +96,6 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   number={1},
   pages={46--55},
   year={1998},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/99.660313},
 }
