+++ paper/paper.bib
@@ -17,7 +17,7 @@
   pages={440701},
   year={2018},
   publisher={Cold Spring Harbor Laboratory},
-  doi={https://doi.org/10.1101/440701},
+  doi={10.1101/440701},
 }
 
 @inproceedings{nickolls2008scalable,
@@ -51,7 +51,7 @@
    month =        {July},
    address =      {Bologna, Italy},
    publisher =    {IEEE},
-   doi={doi.org/10.1109/hpcsim.2014.6903792},
+   doi={10.1109/hpcsim.2014.6903792},
 }
 
 @article{heirendt2019creation,
@@ -74,7 +74,7 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   pages={971--987},
   year={2015},
   publisher={Elsevier},
-  doi={doi: 10.1016/j.cell.2015.05.019},
+  doi={10.1016/j.cell.2015.05.019},
 }
 
 @article{mahadevan2003effects,
