### 2019-03-31 cfe96b2
fixed typo in paper
