+++ paper/paper.bib
@@ -99,3 +99,13 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   publisher={IEEE},
   doi={10.1109/99.660313},
 }
+@article{mahadevan2005applications,
+  title={Applications of metabolic modeling to drive bioprocess development for the production of value-added chemicals},
+  author={Mahadevan, Radhakrishnan and Burgard, Anthony P and Famili, Iman and Van Dien, Steve and Schilling, Christophe H},
+  journal={Biotechnology and bioprocess engineering},
+  volume={10},
+  number={5},
+  pages={408},
+  year={2005},
+  publisher={Springer}
+}
