+++ paper/paper.md
@@ -23,8 +23,7 @@ conditions and environments. It serves as a tool for the support of wet-lab expe
 products, 
 metabolism is the most amenable to modeling because it is directly related to key biological functions and is the support for several drugs targets. 
 Moreover, public data resources of several metabolites and their abundances have been developing rapidly in the recent years. As a biotechnological application, the metabolic modeling 
-of ethanol-producing 
-bacteria allows 
+of ethanol-producing bacteria allows 
 finding key interventions (such as substrate optimization) that would increase the yield in the bioreactor, thereby its efficiency [@mahadevan2005applications].
  
 Recently, high-throughput technologies allowed to generate a large amount of biological data that enabled more complex modeling of biological systems. As models expand in size, the 
@@ -34,7 +33,7 @@ A tool of choice for the analysis of metabolic models is the sampling of the spa
 sampling is an unbiased tool for metabolic modeling that explores all the space of possible metabolic phenotypes. For large models, sampling becomes expensive both in time and 
 computational resources. To make 
 sampling more accessible in the 
-modeler´s toolbox, I present ACHR.cu which is a CUDA-based [@nickolls2008scalable] implementation of the sampling algorithm ACHR [@kaufman1998direction].
+modeler´s toolbox, I present ACHR.cu which is a fast, CUDA-based [@nickolls2008scalable] implementation of the sampling algorithm ACHR [@kaufman1998direction].
 
 # Results
 
@@ -58,52 +57,18 @@ factor to favor the imbalance between the points to generate.
 
 To remediate to this issue, dynamic loading balancing implemented through the OpenMP parallel library in C [@dagum1998openmp] allows assigning fewer points to workers that required 
 more time to solve previous chunks. In the end, the workers converge at the same time.
-Below, I report the run times of the generation of warmup points in MATLAB (CreateWarmupMATLAB) (Table 1) and through a hybrid MPI/OpenMP implementation (CreateWarmupVF) (Table 2), 
-both for the generation of 30,000 warmup points. 
-Since the original implementation in MATLAB does not support parallelism, I reported the run times for the sequential version below. The run times can be divided by the number of cores 
-to get the times (at best) for a parallel run.
-The experiments are the average of three trials in every setting in seconds.
 
-| Model\\Cores  | 1                  |
-|---------------| -------------------|
-| Ecoli core    | 149                |
-| P Putida      | 385                |
-| EcoliK12      | 801                | 
-| Recon2        | 11346              | 
-| E Matrix      | NA*                | 
-| Ec Matrix     | NA*                | 
-
-Table 1: Runtimes of CreateWarmupMATLAB for a set of metabolic models using 1 core. *did not converge after 20,000 seconds.
-
-
-| Model\\Cores  | 1               |2               |4  | 8     | 16| 32|
-|---------------| ----------------|----------------|---|-------|---|---|
-| Ecoli core    | 2.8             |1.8             |0.8|0.7    |0.5|0.5|
-| P Putida      | 12.5            |13              |8  |4      |2  |2  |
-| EcoliK12      | 49              |43              |23 |10.4   |9.5|9.1|
-| Recon2        | 288             |186             |30 |32     |24 |21 |
-| E Matrix      | 602             |508             |130|52     |43 |43 |
-| Ec Matrix     | 5275            |4986            |924|224    |118|117|
-
-Table 2: Runtimes of CreateWarmupVF for a set of metabolic models using 1,2,4,8,16, and 32 cores.
-
-
-The speedup was substantial (up to 50x in some cases) and showed the power of dynamic load balancing in imbalanced metabolic models.
-Also, I noted that the model can be largely imbalanced due to the generation of a random c vector and that averaging three experiments can be insufficient to get the average run time 
[…]
