+++ paper/paper.bib
@@ -107,5 +107,6 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   number={5},
   pages={408},
   year={2005},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/bf02989823}
 }
