### 2019-04-08 #1363 @marouenbg
@lpantano @wmegchel  @prasunanand
I thought that it can be cumbersome to go through all the dependencies with potentially no access to NVIDIA GPU, below I provide a code ocean capsule which is a container with all the dependencies cached. You can also have reproducible runs of an example.
https://codeocean.com/capsule/4395814/tree
I have that helps your assessment.
