### 2019-04-29 #1363 @prasunanand
@marouenbg  I am unable to download
Dependencies
IBM CPLEX v12.6 (free for academics)
