### 2019-04-29 #1363 @marouenbg
@prasunanand Thanks for the feedback. I fixed the missing doi in the summary. Which error did you get when you could not download CPLEX?
Otherwise, you can skip the installation part and check an installed version at code ocean https://codeocean.com/capsule/4395814/tree
