### 2019-04-30 c50a08e
Added source description in README
