### 2019-04-30 55980e8
added make test for createVFwarmup
