+++ paper/paper.md
@@ -10,7 +10,7 @@ authors:
    orcid: 0000-0001-5934-966X
    affiliation: "1"
 affiliations:
- - name: Luxembourg Centre for Systems Biomedicine, University of Luxembourg.
+ - name: Independent researcher.
    index: 1
 date: 03 March 2019
 bibliography: paper.bib
@@ -18,56 +18,47 @@ bibliography: paper.bib
 
 # Summary
 
-The *in silico* modeling of biological organisms consists of the mathematical representation of key functions of a biological system and the study of its behavior in different 
-conditions and environments. It serves as a tool for the support of wet-lab experiments or to generate hypotheses about the functioning of the subsystems. Among the many biological 
+The *in silico* modeling of biological organisms consists of the mathematical representation of key functions of a biological system and the study of its behavior in different conditions and environments. It serves as a tool for the support of wet lab experiments and to generate hypotheses about the functioning of the subsystems. Among the many biological 
 products, 
 metabolism is the most amenable to modeling because it is directly related to key biological functions and is the support for several drugs targets. 
-Moreover, public data resources of several metabolites and their abundances have been developing rapidly in the recent years. As a biotechnological application, the metabolic modeling 
+Moreover, public data resources of several metabolites and their abundances have been developing rapidly in recent years. As a biotechnological application, the metabolic modeling 
 of ethanol-producing bacteria allows 
 finding key interventions (such as substrate optimization) that would increase the yield in the bioreactor, thereby its efficiency [@mahadevan2005applications].
  
-Recently, high-throughput technologies allowed to generate a large amount of biological data that enabled more complex modeling of biological systems. As models expand in size, the 
+Recently, high-throughput technologies allowed to generate a large amount of biological data that enabled complex modeling of biological systems. As models expand in size, the 
 tools used for their analysis have to be appropriately scaled to include the use of parallel software.
 
 A tool of choice for the analysis of metabolic models is the sampling of the space of their possible phenotypes. Instead of considering one specific biological function of interest, 
-sampling is an unbiased tool for metabolic modeling that explores all the space of possible metabolic phenotypes. For large models, sampling becomes expensive both in time and 
-computational resources. To make 
-sampling more accessible in the 
+sampling is an unbiased tool for metabolic modeling that explores all the space of possible metabolic phenotypes. For large models, sampling becomes expensive both in time and computational resources. To make 
+sampling accessible in the 
 modeler´s toolbox, I present ACHR.cu which is a fast, CUDA-based [@nickolls2008scalable] implementation of the sampling algorithm ACHR [@kaufman1998direction].
 
 # Results
 
-Mathematically, metabolic models are networks of metabolites involved in reactions and are formulated as linear programs [@o2015using]. The solution of the linear program allows 
-finding flux distributions in the network that achieve the objective function of interest. ACHR allows obtaining flux distribution for each reaction under the conditions of optimality.
+Metabolic models are networks of metabolites involved in reactions and are formulated as linear programs [@o2015using]. The solution of the linear program allows finding flux distributions in the network that achieve the objective function of interest. ACHR allows obtaining flux distribution for each reaction under the conditions of optimality.
 
 The sampling of the solution space of metabolic models is a two-step process:
 
 ## Generation of warmup points
 
-The first step of sampling the solution space of metabolic networks involves the generation of warmup points that are solutions of the metabolic model's linear program. The sampling 
+The first step of sampling the solution space of metabolic models involves the generation of warmup points that are solutions of the metabolic model's linear program. The sampling 
 MCMC chain starts from those solutions to explore the solution space. The generation of p warmup points corresponds to flux variability analysis (FVA) [@mahadevan2003effects] solutions 
[…]
