+++ paper/paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 The *in silico* modeling of biological organisms consists of the mathematical representation of key functions of a biological system and the study of its behavior in different conditions and environments. It serves as a tool for the support of wet lab experiments and for the generation of hypotheses about the functioning of the subsystems. Among the many biological 
 products, 
 metabolism is the most amenable to modeling because it is directly related to key biological functions and is the support for several drugs targets. 
-Moreover, public data resources of several metabolites and their abundances have been developing rapidly in recent years enabling applications in many areas. In biotechnology, the metabolic modeling 
+Moreover, public data resources of several metabolites and their abundances have been developing rapidly in recent years thereby enabling applications in many areas. In biotechnology, the metabolic modeling 
 of ethanol-producing bacteria allows 
 finding key interventions (such as substrate optimization) that would increase the yield in the bioreactor, thereby its efficiency [@mahadevan2005applications].
  
@@ -35,7 +35,7 @@ modeler´s toolbox, I present ACHR.cu which is a fast, CUDA-based [@nickolls2008
 
 # Results
 
-Metabolic models are networks of m metabolites involved in n reactions and are formulated as linear programs [@o2015using] using the stoichiometric matrix S(m,n). A central concept in metabolic modeling is the steady state assumption, meaning that generating fluxes are equal to degrading fluxes and the rate of change of metabolite concentrations is null. Mathematically, the steady state assumption translates to finding the null space of the stoichiometric matrix S(m,n). 
+Metabolic models are networks of m metabolites involved in n reactions and are formulated as linear programs [@o2015using] using the stoichiometric matrix S(m,n). A central concept in metabolic modeling is the steady state assumption, meaning that generating fluxes are equal to degrading fluxes and the rate of change of metabolite concentrations is null. Mathematically, the steady state assumption translates to constraining the solution space to the null space of the stoichiometric matrix S(m,n). 
 
 Particularly, the solution of the linear program allows finding flux distributions in the network that achieve the objective function of interest. ACHR allows obtaining flux distribution for each reaction under the conditions of optimality.
 
