+++ paper/paper.md
@@ -92,7 +92,7 @@ Finally, ACHR.cu was developed as a high-performance tool for the modeling of me
 The architecture of the parallel GPU implementation of ACHR.cu is similar to the MATLAB Cobra Toolbox [@heirendt2019creation] GpSampler. 
 Another tool, OptGpSampler [@megchelenbrink2014optgpsampler] provides a 40x speedup over GpSampler through a i) C implementation and ii) fewer but longer sampling chains launch.
 Since OptGpSampler performs the generation of the warmup points and the sampling in one process, it is clear from the results of the current work that the speedup achieved with the generation of warmup points is more significant than sampling itself. I decoupled the generation of warmup points from sampling to take advantage of dynamic load balancing with OpenMP. Additionally, in OptGpSampler each worker gets the same amount of points and steps to compute; the problem is statically load balanced by design.
-Conversly, if the generation of warmup points is performed separately from sampling, the problem can be dynamically balanced because the workers can generate an uneven number of warmup points and converge simulatenously. 
+In contrast, if the generation of warmup points is performed separately from sampling, the problem can be dynamically balanced because the workers can generate an uneven number of warmup points and converge simulatenously. 
 
 Finally, future improvements of this work can consider an MPI/CUDA hybrid to take advantage of the multi-GPU architecture of recent NVIDIA cards like the K80. Taken together, the 
 parallel architecture of ACHR.cu allows faster sampling of metabolic models over existing tools thereby enabling the unbiased analyses of large-scale systems biology models.
