+++ paper/paper.md
@@ -46,7 +46,7 @@ The sampling of the solution space of metabolic models is a two-step process:
 The first step of sampling the solution space of metabolic models involves the generation of warmup points that are solutions to the metabolic model's linear program. The sampling 
 Markov Chain Monte Carlo (MCMC) chain starts from those solutions to explore the solution space. 
 
-The generation of p >= 2n warmup points corresponds to flux variability analysis (FVA) [@mahadevan2003effects] solutions 
+The generation of p \geq 2n warmup points corresponds to flux variability analysis (FVA) [@mahadevan2003effects] solutions 
 for the first 2n points, with n the number of reactions in the network, where the objective function is to minimize and maximize each reaction in the model. For the remaining p - 2n points, it corresponds to solutions generated using a random objective vector c(1,n).
 
 The generation of warmup points is a time-consuming process and requires the use of more than one core in parallel. The distribution of the points to generate among the nc cores of the computer is often performed through static balancing with each core getting p/nc points to generate. Nevertheless, the formulation of the problem induces a significant imbalance in the distribution of work, meaning that the workers will not converge at the same time thereby slowing down the overall process. I showed previously that FVA is imbalanced, 
