### 2019-05-24 1b48236
Update ACHRCuda.cu
