+++ paper/paper.bib
@@ -110,3 +110,36 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   publisher={Springer},
   doi={10.1007/bf02989823}
 }
+
+@article{thiele2013community,
+  title={A community-driven global reconstruction of human metabolism},
+  author={Thiele, Ines and Swainston, Neil and Fleming, Ronan MT and Hoppe, Andreas and Sahoo, Swagatika and Aurich, Maike K and Haraldsdottir, Hulda and Mo, Monica L and Rolfsson, Ottar and Stobbe, Miranda D and others},
+  journal={Nature biotechnology},
+  volume={31},
+  number={5},
+  pages={419--425},
+  year={2013},
+  publisher={Nature Research},
+  doi={10.1038/nbt.2488}
+}
+
+@article{orth2010reconstruction,
+  title={Reconstruction and use of microbial metabolic networks: the core Escherichia coli metabolic model as an educational guide},
+  author={Orth, Jeffrey D and Fleming, Ronan MT and Palsson, Bernhard O},
+  journal={EcoSal plus},
+  year={2010},
+  publisher={American Society for Microbiology},
+  doi={10.1128/ecosalplus.10.2.1}
+}
+
+@article{nogales2008genome,
+  title={A genome-scale metabolic reconstruction of Pseudomonas putida KT2440: i JN746 as a cell factory},
+  author={Nogales, Juan and Palsson, Bernhard {\O} and Thiele, Ines},
+  journal={BMC systems biology},
+  volume={2},
+  number={1},
+  pages={79},
+  year={2008},
+  publisher={BioMed Central},
+  doi={10.1186/1752-0509-2-79}
+}
