+++ paper/paper.bib
@@ -143,3 +143,31 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   publisher={BioMed Central},
   doi={10.1186/1752-0509-2-79}
 }
+
+@incollection{li2011gpu,
+  title={A GPU-based parallel algorithm for large scale linear programming problem},
+  author={Li, Jianming and Lv, Renping and Hu, Xiangpei and Jiang, Zhongqiang},
+  booktitle={Intelligent Decision Technologies},
+  pages={37--46},
+  year={2011},
+  publisher={Springer}
+}
+
+@inproceedings{gurung2019simultaneous,
+  title={Simultaneous solving of batched linear programs on a GPU},
+  author={Gurung, Amit and Ray, Rajarshi},
+  booktitle={Proceedings of the 2019 ACM/SPEC International Conference on Performance Engineering},
+  pages={59--66},
+  year={2019},
+  organization={ACM}
+}
+
+@article{charlton2019two,
+  title={Two-dimensional batch linear programming on the GPU},
+  author={Charlton, John and Maddock, Steve and Richmond, Paul},
+  journal={Journal of Parallel and Distributed Computing},
+  volume={126},
+  pages={152--160},
+  year={2019},
+  publisher={Elsevier}
+}
