+++ paper/paper.bib
@@ -150,7 +150,8 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   booktitle={Intelligent Decision Technologies},
   pages={37--46},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-3-642-22194-1_4}
 }
 
 @inproceedings{gurung2019simultaneous,
@@ -159,7 +160,8 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   booktitle={Proceedings of the 2019 ACM/SPEC International Conference on Performance Engineering},
   pages={59--66},
   year={2019},
-  organization={ACM}
+  organization={ACM},
+  doi={10.1145/3297663.3310308}
 }
 
 @article{charlton2019two,
@@ -169,5 +171,6 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   volume={126},
   pages={152--160},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jpdc.2019.01.001}
 }
