### 2019-05-27 #1363 @marouenbg
Hi @lpantano I am all set. A math symbol did not compile in the final text, but it is working now. Thanks to all involved!
