### 2019-05-28 566ce2c
fixed citation format
