### 2019-05-28 506380c
updated maths symbols
