### 2019-05-28 #1363 @lpantano
Hi @openjournals/joss-eics,

I think this is complete!
