### 2019-05-28 de9c5b6
Update paper.bib
Fixes some references
