+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{megchelenbrink2014optgpsampler,
-  title={optGpSampler: an improved tool for uniformly sampling the solution-space of genome-scale metabolic networks},
+  title={{optGpSampler}: an improved tool for uniformly sampling the solution-space of genome-scale metabolic networks},
   author={Megchelenbrink, Wout and Huynen, Martijn and Marchiori, Elena},
   journal={PloS one},
   volume={9},
@@ -44,7 +44,7 @@
 
 @InProceedings{VBCG_HPCS14,
    author =       {S. Varrette and P. Bouvry and H. Cartiaux and F. Georgatos},
-   title =        {Management of an Academic HPC Cluster: The UL Experience},
+   title =        {Management of an Academic {HPC} Cluster: The {UL} Experience},
    booktitle =    {Proc. of the 2014 Intl. Conf. on High Performance Computing \& Simulation (HPCS 2014)},
    year =         {2014},
    pages =        {959--967},
@@ -55,7 +55,7 @@
 }
 
 @article{heirendt2019creation,
-  title={Creation and analysis of biochemical constraint-based models using the COBRA Toolbox v. 3.0},
+  title={Creation and analysis of biochemical constraint-based models using the {COBRA} Toolbox v. 3.0},
   author={Heirendt, Laurent and Arreckx, Sylvain and Pfau, Thomas and Mendoza, Sebasti{\'a}n N and Richelle, Anne and Heinken, Almut and Haraldsd{\'o}ttir, Hulda S and Wachowiak, Jacek 
 and Keating, Sarah M and Vlasov, Vanja and others},
   journal={Nature protocols},
@@ -67,7 +67,7 @@ and Keating, Sarah M and Vlasov, Vanja and others},
 
 @article{o2015using,
   title={Using genome-scale models to predict biological capabilities},
-  author={O’Brien, Edward J and Monk, Jonathan M and Palsson, Bernhard O},
+  author={O'Brien, Edward J and Monk, Jonathan M and Palsson, Bernhard O},
   journal={Cell},
   volume={161},
   number={5},
@@ -90,7 +90,7 @@ and Keating, Sarah M and Vlasov, Vanja and others},
 }
 
 @article{dagum1998openmp,
-  title={OpenMP: An industry-standard API for shared-memory programming},
+  title={{OpenMP}: An industry-standard {API} for shared-memory programming},
   author={Dagum, Leonardo and Menon, Ramesh},
   journal={Computing in Science \& Engineering},
   number={1},
@@ -99,53 +99,56 @@ and Keating, Sarah M and Vlasov, Vanja and others},
   publisher={IEEE},
   doi={10.1109/99.660313},
 }
+
 @article{mahadevan2005applications,
-  title={Applications of metabolic modeling to drive bioprocess development for the production of value-added chemicals},
-  author={Mahadevan, Radhakrishnan and Burgard, Anthony P and Famili, Iman and Van Dien, Steve and Schilling, Christophe H},
-  journal={Biotechnology and bioprocess engineering},
-  volume={10},
-  number={5},
-  pages={408},
-  year={2005},
-  publisher={Springer},
-  doi={10.1007/bf02989823}
+  doi = {10.1007/bf02989823},
+  year = 2005,
+  month = oct,
+  volume = {10},
+  number = {5},
+  pages = {408--417},
+  author = {Radhakrishnan Mahadevan and Anthony P. Burgard and Iman Famili and Steve Van Dien and Christophe H. Schilling},
+  title = {Applications of metabolic modeling to drive bioprocess development for the production of value-added chemicals},
[…]
