### 2019-05-28 #1363 @kyleniemeyer
Hi @marouenbg, I just submitted two minor PRs for the paper and references; could you merge those and then we can rebuild the PDF?
