### 2019-05-28 #1363 @lpantano
Do the last changes change the Zenode DOI or/and version?
