### 2019-05-28 #1363 @kyleniemeyer
@lpantano they were only minor changes to the paper, which we archive, so I wasn't worried about updating the Zenodo record
