### 2019-05-28 #1363 @kyleniemeyer
Congrats @marouenbg on your paper's publication in JOSS! Thanks also to @wmegchel and @prasunanand for reviewing, and @lpantano for editing!
