### 2019-11-25 #1363 @marouenbg
@kyleniemeyer @arfon I just realized I had small typo in the title. First, a dash is missing in `gpu accelerated`. Second the title line ends with a full stop, which gives me references with two full stops. Is it possible to amend the title? Otherwise it is fine like it is.
